+++ cmd/falco/runner.go
@@ -8,11 +8,9 @@ import (
 	"strings"
 	"time"
 
-	"path/filepath"
-
-	"github.com/goccy/go-yaml"
 	"github.com/pkg/errors"
 	"github.com/ysugimoto/falco/ast"
+	"github.com/ysugimoto/falco/config"
 	"github.com/ysugimoto/falco/context"
 	"github.com/ysugimoto/falco/interpreter"
 	icontext "github.com/ysugimoto/falco/interpreter/context"
@@ -26,8 +24,7 @@ import (
 )
 
 var (
-	ErrParser   = fmt.Errorf("parser error")
-	DotfileName = ".falcorc"
+	ErrParser = fmt.Errorf("parser error")
 )
 
 type Level int
@@ -84,7 +81,7 @@ type Runner struct {
 	errors   int
 }
 
-func NewRunner(c *Config, f Fetcher) (*Runner, error) {
+func NewRunner(c *config.Config, f Fetcher) (*Runner, error) {
 	r := &Runner{
 		level:     LevelError,
 		overrides: make(map[string]linter.Severity),
@@ -99,15 +96,13 @@ func NewRunner(c *Config, f Fetcher) (*Runner, error) {
 		// We communicate Fastly API with service id and api key,
 		// lookup fixed environment variable, FASTLY_SERVICE_ID and FASTLY_API_KEY
 		// So user needs to set them with "-r" argument.
-		serviceId := os.Getenv("FASTLY_SERVICE_ID")
-		apiKey := os.Getenv("FASTLY_API_KEY")
-		if serviceId == "" || apiKey == "" {
+		if c.FastlyServiceID == "" || c.FastlyApiKey == "" {
 			return nil, errors.New("Both FASTLY_SERVICE_ID and FASTLY_API_KEY environment variables must be specified")
 		}
 		func() {
 			// Remote communication is optional so we keep processing even if remote communication is failed
 			// We allow each API call to take up to to 5 seconds
-			f := remote.NewFastlyApiFetcher(serviceId, apiKey, 5*time.Second)
+			f := remote.NewFastlyApiFetcher(c.FastlyServiceID, c.FastlyApiKey, 5*time.Second)
 			snippets, err := NewSnippet(f).Fetch()
 			if err != nil {
 				writeln(red, err.Error())
@@ -137,65 +132,29 @@ func NewRunner(c *Config, f Fetcher) (*Runner, error) {
 	}
 
 	// Set verbose level
-	if c.VV {
+	if c.VerboseInfo {
 		r.level = LevelInfo
-	} else if c.V {
+	} else if c.VerboseWarning {
 		r.level = LevelWarning
 	}
 
-	// Make overriding error level setting from rc file
-	r.initOverrides()
+	// Override linter rules
+	for key, value := range c.Rules {
+		switch strings.ToUpper(value) {
+		case "ERROR":
+			r.overrides[key] = linter.ERROR
+		case "WARNING":
+			r.overrides[key] = linter.WARNING
+		case "INFO":
+			r.overrides[key] = linter.INFO
+		case "IGNORE":
+			r.overrides[key] = linter.IGNORE
+		default:
+			writeln(yellow, "level for rule %s has invalid value %s, skip it.", key, value)
+		}
+	}
 
 	return r, nil
-}
-
-func (r *Runner) initOverrides() {
-	// find up rc file
-	cwd, err := os.Getwd()
-	if err != nil {
-		return
-	}
-	var rcFile string
-	for {
-		rcFile = filepath.Join(cwd, DotfileName)
-		if _, err := os.Stat(rcFile); err == nil {
-			break
-		}
-		cwd = filepath.Dir(cwd)
-		if cwd == "/" {
-			// find up to root directory, stop it
-			return
-		}
-	}
-
-	fp, err := os.Open(rcFile)
-	if err != nil {
-		writeln(yellow, "Configuration file found at %s but could not open", rcFile)
-		return
-	}
-	defer fp.Close()
-	o := make(map[string]string)
-	if err := yaml.NewDecoder(fp).Decode(&o); err != nil {
-		writeln(yellow, "Failed to decode configuration file at %s: %s", rcFile, err)
-		return
-	}
-
-	// validate configuration file
-	for k, v := range o {
-		switch strings.ToUpper(v) {
-		case "ERROR":
-			r.overrides[k] = linter.ERROR
-		case "WARNING":
-			r.overrides[k] = linter.WARNING
-		case "INFO":
-			r.overrides[k] = linter.INFO
-		case "IGNORE":
-			r.overrides[k] = linter.IGNORE
-		default:
-			writeln(yellow, "level for rule %s has invalid value %s, skip it.", k, v)
-			return
-		}
-	}
 }
 
 func (r *Runner) Transform(vcl *plugin.VCL) error {