+++ cmd/falco/runner_test.go
@@ -4,6 +4,7 @@ import (
 	"os"
 	"testing"
 
+	"github.com/ysugimoto/falco/config"
 	"github.com/ysugimoto/falco/resolver"
 	"github.com/ysugimoto/falco/terraform"
 )
@@ -27,7 +28,7 @@ func loadFromTfJson(fileName string, t *testing.T) ([]resolver.Resolver, Fetcher
 func TestResolveExternalWithExternalProperties(t *testing.T) {
 	for _, fileName := range []string{"../../terraform/data/terraform-valid.json", "../../terraform/data/terraform-valid-weird-name.json"} {
 		rslv, f := loadFromTfJson(fileName, t)
-		r, err := NewRunner(&Config{V: true}, f)
+		r, err := NewRunner(&config.Config{VerboseWarning: true}, f)
 		if err != nil {
 			t.Fatalf("Unexpected runner creation error: %s", err)
 			return
@@ -51,7 +52,7 @@ func TestResolveExternalWithExternalProperties(t *testing.T) {
 func TestResolveExternalWithNoExternalProperties(t *testing.T) {
 	fileName := "../../terraform/data/terraform-empty.json"
 	rslv, f := loadFromTfJson(fileName, t)
-	r, err := NewRunner(&Config{V: true}, f)
+	r, err := NewRunner(&config.Config{VerboseWarning: true}, f)
 	if err != nil {
 		t.Fatalf("Unexpected runner creation error: %s", err)
 		return
@@ -74,7 +75,7 @@ func TestResolveExternalWithNoExternalProperties(t *testing.T) {
 func TestResolveWithDuplicateDeclarations(t *testing.T) {
 	fileName := "../../terraform/data/terraform-duplicate.json"
 	rslv, f := loadFromTfJson(fileName, t)
-	r, err := NewRunner(&Config{V: true}, f)
+	r, err := NewRunner(&config.Config{VerboseWarning: true}, f)
 	if err != nil {
 		t.Fatalf("Unexpected runner creation error: %s", err)
 	}
@@ -92,7 +93,7 @@ func TestResolveModulesWithVCLExtension(t *testing.T) {
 	fileName := "../../terraform/data/terraform-modules-extension.json"
 	rslv, f := loadFromTfJson(fileName, t)
 
-	r, err := NewRunner(&Config{V: true}, f)
+	r, err := NewRunner(&config.Config{VerboseWarning: true}, f)
 	if err != nil {
 		t.Fatalf("Unexpected runner creation error: %s", err)
 	}
@@ -111,7 +112,7 @@ func TestResolveModulesWithoutVCLExtension(t *testing.T) {
 	fileName := "../../terraform/data/terraform-modules-without-extension.json"
 	rslv, f := loadFromTfJson(fileName, t)
 
-	r, err := NewRunner(&Config{V: true}, f)
+	r, err := NewRunner(&config.Config{VerboseWarning: true}, f)
 	if err != nil {
 		t.Fatalf("Unexpected runner creation error: %s", err)
 	}
@@ -166,7 +167,7 @@ func TestRepositoryExamples(t *testing.T) {
 		},
 	}
 
-	c := &Config{V: true}
+	c := &config.Config{VerboseWarning: true}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			resolvers, err := resolver.NewFileResolvers(tt.fileName, c.IncludePaths)
@@ -174,7 +175,7 @@ func TestRepositoryExamples(t *testing.T) {
 				t.Errorf("Unexpected runner creation error: %s", err)
 				return
 			}
-			r, err := NewRunner(&Config{V: true}, nil)
+			r, err := NewRunner(c, nil)
 			if err != nil {
 				t.Errorf("Unexpected runner creation error: %s", err)
 				return