+++ context/builtin.go
@@ -1816,6 +1816,16 @@ func predefinedVariables() Variables {
 						},
 					},
 				},
+				"host_header": &Object{
+					Items: map[string]*Object{},
+					Value: &Accessor{
+						Get:       types.StringType,
+						Set:       types.NeverType,
+						Unset:     false,
+						Scopes:    RECV | HASH | HIT | MISS | PASS | FETCH | ERROR | DELIVER | LOG,
+						Reference: "https://developer.fastly.com/reference/vcl/variables/client-request/fastly-info-host-header/",
+					},
+				},
 				"is_cluster_edge": &Object{
 					Items: map[string]*Object{},
 					Value: &Accessor{